Move all extensions into their own package
