+++ app/src/main/java/com/hedvig/android/app/DebugFragment.kt
@@ -9,7 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.navigation.findNavController
-import com.hedvig.android.owldroid.util.compatSetTint
+import com.hedvig.android.owldroid.util.extensions.compatSetTint
 import kotlinx.android.synthetic.main.fragment_debug.*
