+++ owldroid/src/main/java/com/hedvig/android/owldroid/ui/profile/ProfileViewModel.kt
@@ -4,7 +4,7 @@ import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
 import com.hedvig.android.owldroid.data.profile.ProfileRepository
 import com.hedvig.android.owldroid.graphql.ProfileQuery
-import com.hedvig.android.owldroid.util.default
+import com.hedvig.android.owldroid.util.extensions.default
 import timber.log.Timber
 import javax.inject.Inject
