+++ owldroid/src/main/java/com/hedvig/android/owldroid/ui/profile/myinfo/MyInfoFragment.kt
@@ -13,7 +13,7 @@ import android.view.*
 import com.hedvig.android.owldroid.R
 import com.hedvig.android.owldroid.di.ViewModelFactory
 import com.hedvig.android.owldroid.ui.profile.ProfileViewModel
-import com.hedvig.android.owldroid.util.*
+import com.hedvig.android.owldroid.util.extensions.*
 import dagger.android.support.AndroidSupportInjection
 import kotlinx.android.synthetic.main.app_bar.*
 import kotlinx.android.synthetic.main.fragment_my_info.*
