+++ owldroid/src/main/java/com/hedvig/android/owldroid/util/extensions/MutableLiveDataExtensions.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.owldroid.util
+package com.hedvig.android.owldroid.util.extensions
 
 import android.arch.lifecycle.MutableLiveData
