+++ owldroid/src/main/java/com/hedvig/android/owldroid/util/extensions/ViewExtensions.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.owldroid.util
+package com.hedvig.android.owldroid.util.extensions
 
 import android.view.View
